- aded fixes and tests for fmu-export and fmu-import
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@11602 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ SimulationRuntime/fmi/import/moGenerator.h
@@ -65,9 +65,9 @@ typedef struct{
 typedef struct{
   const char* declType;
   const char* quantity;
-  long min;
+  int min;
   int defMin;
-  long max;
+  int max;
   int defMax;
   int start;
   int defStart;
